Added support for Non-linear PCA and made corresponding additions to the documentations. Also added simulated example data for use with nlpca.
Changed Q2 cross-validation to guarantee that no full column or row is
ever deleted. This is done by deleting diagonals from the matrix
instead of values with random indices. The drawback here is that no
less than a full diagonal can be deleted so the possible folds are
limited, but this should be of no practical importance.

The screeplot function was removed and replaced with plotR2 as
screeplot is a generic function from stats with a rigid definiton
which can not be safely modified (npcs in argument vs nPcs e.g.). 

Removed exclamation marks from warnings, messaged. I find them
obnoxious. Also, i changed the use of the word 'gene' as a row in an
expression matrix is not a 'gene' but a _variable_ measuring the
expression estimates of a gene.



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/pcaMethods@22378 bc3139a8-67e5-0310-9ffc-ced21a209358